refactor(ops): extract qa-loop command runtime from coding-agent
Refs #933
