+++ crates/tau-ops/Cargo.toml
@@ -9,6 +9,7 @@ anyhow.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
+wait-timeout.workspace = true
 tau-access = { path = "../tau-access" }
 tau-ai = { path = "../tau-ai" }
 tau-cli = { path = "../tau-cli" }
