+++ crates/tau-ops/src/lib.rs
@@ -1,9 +1,11 @@
 mod channel_store_admin;
 mod daemon_runtime;
 mod project_index;
+mod qa_loop_commands;
 mod transport_health;
 
 pub use channel_store_admin::*;
 pub use daemon_runtime::*;
 pub use project_index::*;
+pub use qa_loop_commands::*;
 pub use transport_health::*;
